requirements.txt
# RTX Autonomous Trading System Dependencies

# === CORE DEPENDENCIES ===
yfinance>=0.2.30
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
loguru>=0.7.0
asyncio-mqtt>=0.16.0

# === INTERACTIVE BROKERS ===
ib_insync>=0.9.86

# === AI & MACHINE LEARNING ===
openai>=1.12.0
scikit-learn>=1.3.0

# === TELEGRAM BOT ===
aiohttp>=3.9.0
requests>=2.31.0

# === TECHNICAL ANALYSIS ===
ta-lib>=0.4.25
talib>=0.4.25

# === DATA PROCESSING ===
scipy>=1.11.0
matplotlib>=3.7.0
seaborn>=0.12.0

# === ASYNC SUPPORT ===
aiofiles>=23.0.0
asyncio>=3.4.3

# === WEB FRAMEWORK (for monitoring) ===
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0

# === CLOUD DEPLOYMENT ===
docker>=6.1.0
paramiko>=3.3.0

# === TESTING ===
pytest>=7.4.0
pytest-asyncio>=0.21.0

# === MONITORING & LOGGING ===
prometheus-client>=0.19.0
grafana-api>=1.0.3

# === SECURITY ===
cryptography>=41.0.0
bcrypt>=4.1.0

# === DATABASE (Optional) ===
sqlite3  # Built-in
redis>=5.0.0

# === DEVELOPMENT TOOLS ===
black>=23.0.0
flake8>=6.0.0
mypy>=1.7.0

# === SYSTEM UTILITIES ===
psutil>=5.9.0
schedule>=1.2.0 